@@ -1,5 +1,9 @@
-module.exports = [
-    require('../types/NaN'),
-    require('../types/Infinity'),
-    require('../types/NegativeInfinity')
+import nan from '../types/nan.js';
+import infinity from '../types/infinity.js';
+import NegativeInfinity from '../types/negative-infinity.js';
+
+export default [
+    nan,
+    infinity,
+    NegativeInfinity
 ];
