@@ -1,4 +1,6 @@
-module.exports = [
-    require('./builtin')
+import builtin from './builtin.js';
+
+export default [
+    builtin
     // TODO: Add types that are de-facto universal even though not built-in into ecmasript standard.
 ];
