@@ -1,17 +1,19 @@
-var Typeson = require('typeson');
-exports.Date = [
-    function (x) { return Typeson.toStringTag(x) === 'Date'; },
-    function (date) {
-        var time = date.getTime();
-        if (isNaN(time)) {
-            return 'NaN';
+import Typeson from 'typeson';
+export default {
+    date: {
+        test (x) { return Typeson.toStringTag(x) === 'Date'; },
+        replace (date) {
+            const time = date.getTime();
+            if (isNaN(time)) {
+                return 'NaN';
+            }
+            return time;
+        },
+        revive (time) {
+            if (time === 'NaN') {
+                return new Date(NaN);
+            }
+            return new Date(time);
         }
-        return time;
-    },
-    function (time) {
-        if (time === 'NaN') {
-            return new Date(NaN);
-        }
-        return new Date(time);
     }
-];
+};
