@@ -1,7 +1,8 @@
-var Typeson = require('typeson');
-var makeArray = require('../utils/array-from-iterator');
-exports.Set = [
-    function (x) { return Typeson.toStringTag(x) === 'Set'; },
-    function (set) { return makeArray(set.values()); },
-    function (values) { return new Set(values); }
-];
+import Typeson from 'typeson';
+export default {
+    set: {
+        test (x) { return Typeson.toStringTag(x) === 'Set'; },
+        replace (set) { return Array.from(set.values()); },
+        revive (values) { return new Set(values); }
+    }
+};
