@@ -1,32 +1,43 @@
-var Typeson = require('typeson');
-var _global = typeof self === 'undefined' ? global : self;
+/* eslint-env browser, node */
+import Typeson from 'typeson';
+const _global = typeof self === 'undefined' ? global : self;
 
 // Support all kinds of typed arrays (views of ArrayBuffers)
+const exportObj = {};
 [
-    "Int8Array",
-    "Uint8Array",
-    "Uint8ClampedArray",
-    "Int16Array",
-    "Uint16Array",
-    "Int32Array",
-    "Uint32Array",
-    "Float32Array",
-    "Float64Array"
+    'Int8Array',
+    'Uint8Array',
+    'Uint8ClampedArray',
+    'Int16Array',
+    'Uint16Array',
+    'Int32Array',
+    'Uint32Array',
+    'Float32Array',
+    'Float64Array'
 ].forEach(function (typeName) {
-    var TypedArray = _global[typeName];
-    if (TypedArray) exports[typeName] = [
-        function test (x) { return Typeson.toStringTag(x) === typeName; },
-        function encapsulate (a) { return (a.byteOffset === 0 && a.byteLength === a.buffer.byteLength ?
-            // socket.io supports streaming ArrayBuffers. If we have a typed array
-            // representing a portion of the buffer, we need to clone the buffer before leaving it
-            // to socket.io.
-            a : a.slice(0)).buffer;
-        },
-        function revive (buf) {
-            // One may configure socket.io to revive binary data as Buffer or Blob.
-            // We should therefore not rely on that the instance we get here is an ArrayBuffer
-            // If not, let's assume user wants to receive it as configured with socket.io.
-            return Typeson.toStringTag(buf) === 'ArrayBuffer' ? new TypedArray(buf) : buf;
-        }
-    ];
+    const arrType = typeName;
+    const TypedArray = _global[typeName];
+    if (TypedArray) {
+        exportObj[typeName.toLowerCase()] = {
+            test (x) { return Typeson.toStringTag(x) === arrType; },
+            replace (a) {
+                return (a.byteOffset === 0 && a.byteLength === a.buffer.byteLength
+                    ? a
+                    // socket.io supports streaming ArrayBuffers. If we have a typed array
+                    // representing a portion of the buffer, we need to clone the buffer before leaving it
+                    // to socket.io.
+                    : a.slice(0)).buffer;
+            },
+            revive (buf) {
+                // One may configure socket.io to revive binary data as Buffer or Blob.
+                // We should therefore not rely on that the instance we get here is an ArrayBuffer
+                // If not, let's assume user wants to receive it as configured with socket.io.
+                return Typeson.toStringTag(buf) === 'ArrayBuffer'
+                    ? new TypedArray(buf)
+                    : buf;
+            }
+        };
+    }
 });
+
+export default exportObj;
