@@ -5,20 +5,32 @@
    The Function type is not included because their closures would not be serialized, so a revived
    Function that uses closures would not behave as expected. Symbols are similarly not included.
 */
-module.exports = [
+
+import presetUndefined from '../presets/undef.js';
+import primitiveObjects from '../types/primitive-objects.js';
+import specialNumbers from '../presets/special-numbers.js';
+import date from '../types/date.js';
+import error from '../types/error.js';
+import errors from '../types/errors.js';
+import regexp from '../types/regexp.js';
+import map from '../types/map.js';
+import set from '../types/set.js';
+import arraybuffer from '../types/arraybuffer.js';
+import typedArrays from '../types/typed-arrays.js';
+import dataview from '../types/dataview.js';
+import intlTypes from '../types/intl-types.js';
+
+const expObj = [
     // ES5
-    require('../presets/undefined'),
-    require('../types/primitive-objects'),
-    require('../presets/special-numbers'),
-    require('../types/date'),
-    require('../types/error'),
-    require('../types/errors'),
-    require('../types/regexp'),
+    presetUndefined, primitiveObjects, specialNumbers,
+    date, error, errors, regexp
+].concat(
     // ES2015 (ES6)
-    typeof Map === 'function' && require('../types/map'),
-    typeof Set === 'function' && require('../types/set'),
-    typeof ArrayBuffer === 'function' && require('../types/arraybuffer'),
-    typeof Uint8Array === 'function' && require('../types/typed-arrays'),
-    typeof DataView === 'function' && require('../types/dataview'),
-    typeof Intl !== 'undefined' && require('../types/intl-types')
-];
+    typeof Map === 'function' ? map : [],
+    typeof Set === 'function' ? set : [],
+    typeof ArrayBuffer === 'function' ? arraybuffer : [],
+    typeof Uint8Array === 'function' ? typedArrays : [],
+    typeof DataView === 'function' ? dataview : [],
+    typeof Intl !== 'undefined' ? intlTypes : []
+);
+export default expObj;
