@@ -1,10 +1,36 @@
+/* eslint-env node */
+'use strict';
+const ns = require('node-static');
+const opn = require('opn');
+const url = require('url');
+
+const portIdx = 3;
+const defaultPort = 8085;
+const port = parseInt(process.argv[portIdx] || defaultPort, 10);
+
+const file = new ns.Server(undefined, {cache: false});
+
 require('http').createServer(function (req, res) {
-    var extra = req.url === '/browser-test/' ? 'index.html' : '';
-    if ((/\.css$/).test(req.url)) {
-        res.setHeader('Content-Type', 'text/css');
-    }
-    var s = require('fs').createReadStream(((/\.\.\//).test(req.url) ? '' : '.') + req.url + extra);
-    s.pipe(res);
-    s.on('error', function () {});
-}).listen(8085);
-console.log('Started server; open http://localhost:8085/browser-test/index.html or http://localhost:8085/browser-test/worker.html in the browser');
+    req.addListener('end', function () {
+        file.serve(req, res);
+    }).resume();
+}).listen(port);
+
+const siteIdx = 2;
+const defaultURL = 'http://127.0.0.1/';
+let site = process.argv[siteIdx]
+    ? `http://127.0.0.1:${port}/` + process.argv[siteIdx]
+    : defaultURL;
+
+if (port !== defaultPort) {
+    const portAdded = url.parse(site);
+    portAdded.host = portAdded.host.replace(/:.*$/, '') + ':' + port;
+    site = url.format(portAdded);
+}
+
+if (site === '0') {
+    console.log('Started server; open a file within ' + defaultURL + ':' + port + ' in the browser. Tests may take a while to load!'); // eslint-disable-line no-console
+} else {
+    console.log('Started server; opening ' + site + ' in the browser.' + (site.match(/test/) ? ' May take a while to load!' : '')); // eslint-disable-line no-console
+    opn(site);
+}
