@@ -1,36 +1,41 @@
-var Typeson = require('typeson');
-exports.Blob = {
-    test: function (x) { return Typeson.toStringTag(x) === 'Blob'; },
-    replace: function encapsulate (b) { // Sync
-        var req = new XMLHttpRequest();
-        req.open('GET', URL.createObjectURL(b), false); // Sync
-        if (req.status !== 200 && req.status !== 0) throw new Error('Bad Blob access: ' + req.status);
-        req.send();
-        return {
-            type: b.type,
-            stringContents: req.responseText
-        };
-    },
-    revive: function (o) {return new Blob([o.stringContents], {
-        type: o.type
-    });},
-    replaceAsync: function encapsulateAsync (b) {
-        return new Typeson.Promise(function (resolve, reject) {
-            if (b.isClosed) { // On MDN, but not in https://w3c.github.io/FileAPI/#dfn-Blob
-                reject(new Error('The Blob is closed'));
-                return;
+import Typeson from 'typeson';
+
+export default {
+    blob: {
+        test (x) { return Typeson.toStringTag(x) === 'Blob'; },
+        replace (b) { // Sync
+            const req = new XMLHttpRequest();
+            req.open('GET', URL.createObjectURL(b), false); // Sync
+            if (req.status !== 200 && req.status !== 0) {
+                throw new Error('Bad Blob access: ' + req.status);
             }
-            var reader = new FileReader();
-            reader.addEventListener('load', function () {
-                resolve({
-                    type: b.type,
-                    stringContents: reader.result
+            req.send();
+            return {
+                type: b.type,
+                stringContents: req.responseText
+            };
+        },
+        revive ({type, stringContents}) {
+            return new Blob([stringContents], {type});
+        },
+        replaceAsync (b) {
+            return new Typeson.Promise((resolve, reject) => {
+                if (b.isClosed) { // On MDN, but not in https://w3c.github.io/FileAPI/#dfn-Blob
+                    reject(new Error('The Blob is closed'));
+                    return;
+                }
+                const reader = new FileReader();
+                reader.addEventListener('load', () => {
+                    resolve({
+                        type: b.type,
+                        stringContents: reader.result
+                    });
                 });
+                reader.addEventListener('error', () => {
+                    reject(reader.error);
+                });
+                reader.readAsText(b);
             });
-            reader.addEventListener('error', function () {
-                reject(reader.error);
-            });
-            reader.readAsText(b);
-        });
+        }
     }
 };
