@@ -1,41 +1,48 @@
-var Typeson = require('typeson');
-exports.File = {
-    test: function (x) { return Typeson.toStringTag(x) === 'File'; },
-    replace: function encapsulate (f) { // Sync
-        var req = new XMLHttpRequest();
-        req.open('GET', URL.createObjectURL(f), false); // Sync
-        if (req.status !== 200 && req.status !== 0) throw new Error('Bad Blob access: ' + req.status);
-        req.send();
-        return {
-            type: f.type,
-            stringContents: req.responseText,
-            name: f.name,
-            lastModified: f.lastModified
-        };
-    },
-    revive: function (o) {return new File([o.stringContents], o.name, {
-        type: o.type,
-        lastModified: o.lastModified
-    });},
-    replaceAsync: function encapsulateAsync (f) {
-        return new Typeson.Promise(function (resolve, reject) {
-            if (f.isClosed) { // On MDN, but not in https://w3c.github.io/FileAPI/#dfn-Blob
-                reject(new Error('The File is closed'));
-                return;
+import Typeson from 'typeson';
+
+export default {
+    file: {
+        test (x) { return Typeson.toStringTag(x) === 'File'; },
+        replace (f) { // Sync
+            const req = new XMLHttpRequest();
+            req.open('GET', URL.createObjectURL(f), false); // Sync
+            if (req.status !== 200 && req.status !== 0) {
+                throw new Error('Bad Blob access: ' + req.status);
             }
-            var reader = new FileReader();
-            reader.addEventListener('load', function () {
-                resolve({
-                    type: f.type,
-                    stringContents: reader.result,
-                    name: f.name,
-                    lastModified: f.lastModified
+            req.send();
+            return {
+                type: f.type,
+                stringContents: req.responseText,
+                name: f.name,
+                lastModified: f.lastModified
+            };
+        },
+        revive ({name, type, stringContents, lastModified}) {
+            return new File([stringContents], name, {
+                type,
+                lastModified
+            });
+        },
+        replaceAsync (f) {
+            return new Typeson.Promise(function (resolve, reject) {
+                if (f.isClosed) { // On MDN, but not in https://w3c.github.io/FileAPI/#dfn-Blob
+                    reject(new Error('The File is closed'));
+                    return;
+                }
+                const reader = new FileReader();
+                reader.addEventListener('load', function () {
+                    resolve({
+                        type: f.type,
+                        stringContents: reader.result,
+                        name: f.name,
+                        lastModified: f.lastModified
+                    });
                 });
+                reader.addEventListener('error', function () {
+                    reject(reader.error);
+                });
+                reader.readAsText(f);
             });
-            reader.addEventListener('error', function () {
-                reject(reader.error);
-            });
-            reader.readAsText(f);
-        });
+        }
     }
 };
